requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
MarkupSafe==1.1.1
Send2Trash==1.5.0 
altair==4.1.0 
argon2-cffi==20.1.0 
astor==0.8.1 
async-generator==1.10 
attrs==20.2.0 
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.1
bleach==3.2.1
blinker==1.4
blis==0.4.1 
boto3==1.15.1 
botocore==1.18.1 
bottleneck==1.3.2 
cachetools==4.1.1 
catalogue==1.0.0 
certifi==2020.6.20 
cffi==1.14.2 
chardet==3.0.4 
click==7.1.2 
cycler==0.10.0 
cymem==2.0.3 
dataclasses
decorator==4.4.2 
defusedxml==0.6.0 
entrypoints==0.3 
enum-compat==0.0.3 
fastai==1.0.61
Flask==2.0.1
fastprogress==1.0.0 
future==0.18.2 
idna==2.10 
importlib-metadata==1.7.0 
ipykernel==5.3.4 
ipython==7.16.1 
ipython-genutils==0.2.0 
ipywidgets==7.5.1 
jedi==0.17.2 
jinja2==2.11.2 
jmespath==0.10.0 
jsonschema==3.2.0 
jupyter-client==6.1.7 
jupyter-core==4.6.3 
jupyterlab-pygments==0.1.1 
kiwisolver==1.2.0 
matplotlib==3.3.2 
mistune==0.8.4 
murmurhash==1.0.2 
nbclient==0.5.0 
nbconvert==6.0.4 
nbformat==5.0.7 
nest-asyncio==1.4.0 
notebook==6.1.4 
numexpr==2.7.1 
numpy==1.17.2 
nvidia-ml-py3==7.352.0 
packaging==20.4 
pandas==1.1.2 
pandocfilters==1.4.2 
parso==0.7.1 
pathtools==0.1.2 
pexpect==4.8.0 
pickleshare==0.7.5 
pillow==7.2.0 
plac==1.1.3 
preshed==3.0.2 
prometheus-client==0.8.0 
prompt-toolkit==3.0.7 
protobuf==3.13.0 
ptyprocess==0.6.0 
pyarrow==1.0.1 
pycparser==2.20 
pydeck==0.5.0b1 
pygments==2.7.1 
pyparsing==2.4.7 
pyrsistent==0.17.3 
python-dateutil==2.8.1 
pytz==2020.1 
pyyaml==5.3.1 
pyzmq==19.0.2 
requests==2.24.0 
s3transfer==0.3.3 
scipy==1.5.2 
six==1.15.0 
soupsieve==2.0.1 
spacy==2.3.2 
srsly==1.0.2 
streamlit==0.64.0 
terminado==0.9.0 
testpath==0.4.4 
thinc==7.4.1 
toml==0.10.1 
toolz==0.10.0 
torch==1.6.0+cpu
torchvision===0.7.0+cpu
tornado==6.0.4 
tqdm==4.49.0 
traitlets==4.3.3 
tzlocal==2.1 
urllib3==1.25.10 
validators==0.18.1 
wasabi==0.8.0 
watchdog==0.10.3 
wcwidth==0.2.5 
webencodings==0.5.1 
widgetsnbextension==3.5.1 
zipp==3.1.0